requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
pip==24.2
psycopg2-binary==2.9.9
setuptools==72.1.0
sqlparse==0.5.1
wheel==0.43.0